+++ app/src/main/java/dev/leonlatsch/photok/gallery/albums/domain/model/AlbumPhotoRef.kt
@@ -19,5 +19,5 @@ package dev.leonlatsch.photok.gallery.albums.domain.model
 data class AlbumPhotoRef(
     val albumUUID: String,
     val photoUUID: String,
-    val linkedAt: Int,
+    val linkedAt: Long,
 )
